requirements.txt
pandas==1.1.5
PyYAML==6.0.1
requests==2.27.1
SQLAlchemy==1.4.49